{{cookiecutter.github_repository_name}}/requirements.txt
# Core
pytz==2018.9
Django==2.1.5
django-configurations==2.1
gunicorn==19.9.0
newrelic==4.12.0.113

# Auth
django-allauth==0.38.0
django-rest-auth==0.9.3
argon2-cffi==19.1.0

# For the persistence stores
psycopg2-binary==2.7.7
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
Markdown==3.0.1
django-filter==2.1.0
djangorestframework-filters==0.10.2.post0
drf-writable-nested 0.5.1

# Developer Tools
ipdb==0.11
ipython==7.2.0
mkdocs==1.0.4
flake8==3.7.3
django-extensions==2.1.5
pre-commit==1.14.3
black==18.9b0

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.6
nose-progressive==1.5.2
coverage==4.5.2
pytest==4.2.0
pytest-django==3.4.7
pytest-cov==2.6.1

# Static and Media Storage
django-storages==1.7.1
boto3==1.9.86

# Webpack / Angular2 Stuff
django-webpack-loader==0.6.0

# Phone / Address
django-phonenumber-field==2.2.0
phonenumbers==8.10.4

# Date / Time
python-dateutil==2.8.0
